requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
colorclass==2.2.2
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.1
django-countries==7.2.1
django-phonenumber-field==6.0.0
docopt==0.6.2
filelock==3.4.2
idna==3.3
packaging==21.3
pbr==5.8.0
phonenumbers==8.12.40
pip-upgrader==1.4.15
platformdirs==2.4.1
pyparsing==3.0.6
requests==2.27.0
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
terminaltables==3.1.10
tzdata==2021.5
urllib3==1.26.7
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
whitenoise==5.3.0
